mas persistencia

+++ Presentacion/frmRegistroLote.cs
@@ -47,7 +47,7 @@ namespace Presentacion {
                     ClsNdispositivo Ne = new ClsNdispositivo();
                     Ne.MtdGuardarDispositivo(E);
                     //para guardar kardex
-                    ClsEkardex objEKardex = ClsEkardex.crear(txtCodigo.Text, frmAdministrador.data.Rows[0][0].ToString(), "ENTRADA", Convert.ToInt32(txtCantidad.Text), Convert.ToDouble(txtPrecio.Text), "1", DateTime.Now.ToShortTimeString(), Convert.ToDateTime(DateTime.Now.ToShortDateString()));
+                    ClsEkardex objEKardex = ClsEkardex.crear(txtCodigo.Text, frmAdministrador.data.Rows[0][0].ToString(), "ENTRADA", Convert.ToInt32(txtCantidad.Text), Convert.ToDecimal(txtPrecio.Text), "1", DateTime.Now.ToShortTimeString(), Convert.ToDateTime(DateTime.Now.ToShortDateString()));
                     ClsNcomprobante objN = new ClsNcomprobante();
                     objN.MtdAgregarKardex(objEKardex, "ENTRADA");
